Migrate 'contains' test
